requirements.txt
django
psycopg2
djangorestframework
django-cron
requests
cryptography
